requirements.txt
requests
pyyaml
sqlalchemy
pypdf
pdfplumber
jupyter